requirements.txt
fastapi[standard]==0.112.0
requests
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pyyml
pandas